Sidebar and dashboard style
Dashboard and sidebar with basic style.
Logout button added to the sidebar, its very simple for now.
Token of the user saved on the login component mfor further implementation on navigation trough the pages. Only the users with token can go to the different pages -> still needs to be implemented.